Make minor adjustment to /comments POST route
The new comment was being saved in a Promise.all call alongside the
comment's parent post having its 'comments' array updated with the new
comment's _id, but this means the comment could still be saved if
something goes wrong when updating the parent post. Fixed by making the
comment save ONLY if the parent post is successfully found.