Create JWT authorisation for protected routes
Wrote a utility function that handles the authentication of
JSONWebTokens in the 'authorization' header of the request object to
make sure the token is valid. Also checks that the User is an admin; so
the function can just be added at the start of the middleware chain for
any protected routes.