Move responsibility for fetching comment into Comment component
Removed the 'populate' nesting chain on the server-side query for a
post's comment and instead have just passed the responsibility for
fetching a comment's information to the Comment component. When the
comment information is returned by the server, the Comment component can
then render additional child Comment components for each reply, up to
a maximum specified level of nesting.