Change how comments are deleted
Instead of deleting the record in the database, change a 'deleted'
field to true on the record. Also no longer recursively delete replies
to comments on comment deletion. Deleted comments will now be returned
by the server with their name and text fields as "Deleted", but their
replies will still be returned as normal.